Change param type (visualizer)

+++ include/visualizer/visualizer.h
@@ -85,7 +85,7 @@ private:
     int ViewNumber;
     int EdgeMapSaveNumber;
     int DF_ThreshMatchNum;
-    int DF_ThreshRelRho;
+    double DF_ThreshRelRho;
     int DF_IterNum;
 
     int XLibView;
